create: role and permission authorization and email send

+++ app/Http/Controllers/Api/v1/AssetApiController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers\Api\v1;
 
 use App\Http\Controllers\Controller;
 use App\Models\Asset;
+use App\Services\FilterQueryService;
 use App\Traits\ResponserTrait;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class AssetApiController extends Controller
 {
@@ -17,6 +20,8 @@ class AssetApiController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('asset_access'), $this->respondPermissionDenied());
+
         $assets = Asset::paginate(10);
 
         return $this->respondCollectionWithPagination('success', $assets);
@@ -31,6 +36,8 @@ class AssetApiController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('asset_create'), $this->respondPermissionDenied());
+
         Asset::create([
             'asset_name' => $request->asset_name,
             'asset_type_id' => $request->asset_type_id,
@@ -57,8 +64,10 @@ class AssetApiController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request,Asset $asset)
+    public function update(Request $request, Asset $asset)
     {
+        abort_if(Gate::denies('asset_update'), $this->respondPermissionDenied());
+
         $asset->update([
             'asset_name' => $request->asset_name,
             'asset_type_id' => $request->asset_type_id,
@@ -75,8 +84,21 @@ class AssetApiController extends Controller
      */
     public function destroy(Asset $asset)
     {
+        abort_if(Gate::denies('asset_delete'), $this->respondPermissionDenied());
+
         $asset->delete();
 
         return $this->respondSuccessMsgOnly('success');
+    }
+
+    public function query(Request $request)
+    {
+        abort_if(Gate::denies('asset_query'), $this->respondPermissionDenied());
+
+        $assets = DB::table('assets');
+
+        $data = FilterQueryService::FilterQuery($request, $assets);
+
+        return $this->respondCollectionWithPagination('success', $data);
     }
 }
